move lti13 config to handlers folder
